requirements.txt
-e git+https://github.com/thu-coai/AISafetyLab.git@00b582d978c9cef7d11c49ac078a35e9836c4e10#egg=aisafetylab
-e git+https://github.com/EleutherAI/lm-evaluation-harness.git@4c26a9c176ecfb40b369503ce211e356bb800fdc#egg=lm_eval
alpaca-eval
accelerate==1.2.1
anthropic==0.31.0
blake3==1.0.4
bleach==6.1.0
cmake==3.26.3
common==0.1.2
datasets==3.3.2
deepspeed==0.15.3
fire==0.5.0
huggingface-hub
ipython==8.12.3
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
jiter==0.5.0
jsonlines==4.0.0
mamba==1.4.2
matplotlib==3.8.2
ninja==1.11.1.1
nltk==3.6.3
numba==0.60.0
openai==1.64.0
openpyxl==3.1.2
pandas==2.0.3
peft==0.12.0
pillow==11.1.0
pip==24.3.1
PyYAML==6.0.1
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.5.2
scipy==1.12.0
seaborn==0.13.0
setuptools==69.1.0
sympy==1.13.1
tensorboard==2.19.0
tokenizers==0.21.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
triton==3.1.0
trl==0.15.1
vllm==0.7.3
wandb==0.15.12
wasabi==1.1.3
wheel==0.38.4
